# openapi.yaml
openapi: 3.1.0
info:
  title: NovaMember API
  version: 1.0.0
  description: |
    Shopify + Recharge membership management platform with AI-powered engagement scoring.
    All authenticated endpoints require a Bearer JWT obtained from `/api/auth/login` or `/api/checkout`.
  contact:
    name: NovaMember Support
    email: support@novamember.io
  license:
    name: Proprietary
servers:
  - url: https://novamember.io
    description: Production
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/subscriptions:
    get:
      tags: [Subscriptions]
      summary: List member subscriptions
      description: Returns DB records merged with live Recharge data (status, charges).
      responses:
        '200':
          description: Subscription list
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscriptions:
                    type: array
                    items: { $ref: '#/components/schemas/Subscription' }
                  rechargeData: { type: object }
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags: [Subscriptions]
      summary: Cancel a subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subscriptionId]
              properties:
                subscriptionId:    { type: string }
                cancellationReason: { type: string }
      responses:
        '200':
          description: Cancelled
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Subscription not found

    patch:
      tags: [Subscriptions]
      summary: Pause or resume a subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subscriptionId, action]
              properties:
                subscriptionId: { type: string }
                action:         { type: string, enum: [pause, resume] }
      responses:
        '200':
          description: Updated
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/evals:
    post:
      tags: [Evals]
      summary: Run AI engagement scoring pipeline (cron)
      description: |
        Protected by `x-cron-secret` header. Selects members due for re-score
        (lastScoredAt null or >24h), calls Grok (grok-2-1212) for each, persists scores,
        updates engagement tiers, and flags COLD members for at-risk outreach.
      security:
        - cronSecret: []
      responses:
        '200':
          description: Eval run summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  evaluated:       { type: integer }
                  tierDistribution: { type: object }
                  tokenCost:       { type: object }
                  durationMs:      { type: integer }

    get:
      tags: [Evals]
      summary: Last eval run summary + tier distribution
      responses:
        '200':
          description: Summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  lastRun:   { type: object }
                  tiers:     { type: object }
                  coldCount: { type: integer }

  # ── Compliance ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 15-minute access JWT from `/api/auth/login` or `/api/checkout`
    cronSecret:
      type: apiKey
      in: header
      name: x-cron-secret
      description: Vercel cron secret from `CRON_SECRET` env var
    rechargeHmac:
      type: apiKey
      in: header
      name: X-Recharge-Hmac-Sha256
    shopifyHmac:
      type: apiKey
      in: header
      name: X-Shopify-Hmac-Sha256
  responses:
    Unauthorized:
      description: Authentication required or token expired
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error:   { type: string }
        code:    { type: string, enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, RATE_LIMITED, EXTERNAL_SERVICE_ERROR, PAYMENT_REQUIRED, INTERNAL_ERROR] }
        details: { type: object, nullable: true }
...
    Subscription:
      type: object
      properties:
        id:                     { type: string }
        plan:                   { type: string }
        price:                  { type: number }
        billingCycle:           { type: string }
        status:                 { type: string, enum: [ACTIVE, CANCELLED, PAUSED, PAST_DUE] }
        currentPeriodEnd:       { type: string, format: date-time }
        rechargeSubscriptionId: { type: string, nullable: true }
...